refactor: props와 style 타입 분리
